Verified all include/*/*.h header files are clean as far as `spling -strict` is concerned.

+++ elks/arch/i86/lib/bitops.c
@@ -7,7 +7,7 @@
  *	Messy as we lack atomic bit operations on an 8086.
  */
  
-unsigned char clear_bit(unsigned int bit,void *addr)
+unsigned char clear_bit(unsigned short int bit,void *addr)
 {
     unsigned char *ptr = addr;
     unsigned int r, mask, offset = (bit / 8);
@@ -23,7 +23,7 @@ unsigned char clear_bit(unsigned int bit,void *addr)
     return (r ? 1:0);
 }
 
-unsigned char set_bit(unsigned int bit,void *addr)
+unsigned char set_bit(unsigned short int bit,void *addr)
 {
     unsigned char *ptr = addr;
     unsigned int r, mask, offset = (bit / 8);
@@ -39,7 +39,7 @@ unsigned char set_bit(unsigned int bit,void *addr)
     return (r ? 1 : 0);
 }
 
-unsigned char test_bit(unsigned int bit,void *addr)
+unsigned char test_bit(unsigned short int bit,void *addr)
 {
     unsigned char *ptr = addr;
     unsigned int mask, offset = bit / 8;
@@ -56,9 +56,9 @@ unsigned char test_bit(unsigned int bit,void *addr)
  */
 
 /* Use the old faithful version */
-unsigned int find_first_non_zero_bit(void *addr,unsigned int len)
+unsigned short int find_first_non_zero_bit(void *addr, unsigned short int len)
 {
-    unsigned int i;
+    unsigned short int i;
 
     for (i = 0; i < len; i++)
 	if (test_bit(i, addr))
@@ -67,9 +67,9 @@ unsigned int find_first_non_zero_bit(void *addr,unsigned int len)
 }
 
 /* Use the old faithful version */
-unsigned int find_first_zero_bit(void *addr,unsigned int len)
+unsigned short int find_first_zero_bit(void *addr,unsigned short int len)
 {
-    unsigned int i;
+    unsigned short int i;
 
     for (i = 0; i < len; i++)
 	if (!test_bit(i, addr))
